moved stuff in roomrequest file to the model in room_request
